extract LinkCard component
