+++ packages/app/src/lib/components/content/thread/LinkArea.svelte
@@ -18,7 +18,7 @@
   import { decodeTime } from "ulidx";
   import { onNavigate } from "$app/navigation";
   import { getLinkEmbedData } from "$lib/utils/getLinkEmbedData";
-
+  import LinkCard from "./message/LinkCard.svelte";
   export type Message = {
     id: string;
     content: string;
@@ -219,8 +219,6 @@
       setTimeout(() => (isShifting = false), 1000);
     }
   });
-  const cr = (value: string | undefined, seperator = ""): string =>
-    value !== undefined ? ` ${seperator} ${value}` : "";
 </script>
 
 <div class="grow min-h-0 relative">
@@ -288,75 +286,13 @@
             >
               {#snippet children(message: Message)}
                 {#await getLinkEmbedData(message.link)}
-                  waiting
-                {:then d}
-                  {#if d}
-                    <div
-                      class="flex flex-col flex-wrap justify-stretch gap-4 min-[500px]:flex-row"
-                    >
-                      <div class="min-w-0 flex-1 px-3 py-2 flex flex-col">
-                        <p class="mb-1 mt-0 text-sm leading-none opacity-70">
-                          {cr(d.p?.n)}
-                          {cr(d.au?.n, "-")}
-                        </p>
-                        <p class="mb-1 mt-1 line-clamp-2 leading-snug">
-                          <b class="font-bold">{d.t}</b>
-                        </p>
-                        <p class="my-0 line-clamp-4 text-sm leading-tight">
-                          {cr(d.d)}
-                        </p>
-                        <div class="grow py-2"></div>
-                        {#if d.footer}
-                          <p class="mt-2 mb-0 text-sm">{d.footer.t}</p>
-                        {/if}
-                        <a href={message.link} class="title">
-                          <div
-                            class="text-sm leading-tight underline text-blue-400"
-                          >
-                            {message.link}
-                          </div>
-                        </a>
-                      </div>
-                      {#if d.imgs && d.imgs.length}
-                        <div
-                          class=" w-full flex-shrink-0 p-2 min-[500px]:max-w-40"
-                        >
-                          <img
-                            alt={d.imgs[0]?.d ?? ""}
-                            class="my-0 h-full w-full rounded object-cover"
-                            src={d.imgs[0]?.u}
-                          />
-                        </div>
-                      {/if}
-                      ${#if d.vid}
-                        <div
-                          class=" w-full flex-shrink-0 p-2 min-[500px]:max-w-40"
-                        >
-                          <video
-                            class="my-0 h-full w-full rounded object-cover"
-                            poster={d.thumb?.u}
-                            src={d.vid.u}
-                          >
-                            <track kind="captions" />
-                          </video>
-                        </div>
-                      {/if}
-                    </div>
-
-                    <div
-                      class="prose prose-a:text-accent-600 dark:prose-a:text-accent-400"
-                    >
-                      <div>img</div>
-                      <a
-                        href={message.link}
-                        target="_blank"
-                        class="text-link hover:underline"
-                        >{message.link}
-                      </a>
-                    </div>
-                  {:else}
-                    failed
-                  {/if}
+                    waiting
+                  {:then embed}
+                    {#if embed}
+                      <LinkCard {embed} url={message.link.url} />
+                    {:else}
+                      failed
+                    {/if}
                 {/await}
               {/snippet}
             </Virtualizer>
